feat: Separate the ticket document from the ticket model.
